Move client files to new client folder
